jobs/etl_jobs/external/brevo_refacto/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via aiohttp
brevo-python==1.1.2
    # via -r requirements.in
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via
    #   brevo-python
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via typer
db-dtypes==1.4.3
    # via pandas-gbq
dotenv==0.9.9
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.25.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   pandas-gbq
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   pandas-gbq
    #   pydata-google-auth
google-auth-oauthlib==0.7.0
    # via
    #   -r requirements.in
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-bigquery==3.35.0
    # via
    #   -r requirements.in
    #   pandas-gbq
google-cloud-core==2.4.3
    # via google-cloud-bigquery
google-cloud-secret-manager==2.20.0
    # via -r requirements.in
google-crc32c==1.7.1
    # via google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-bigquery
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-secret-manager
grpcio==1.73.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
idna==3.10
    # via
    #   requests
    #   yarl
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
numpy==1.26.4
    # via
    #   -r requirements.in
    #   db-dtypes
    #   pandas
    #   pandas-gbq
    #   pyarrow
oauthlib==3.3.1
    # via requests-oauthlib
packaging==25.0
    # via
    #   db-dtypes
    #   google-cloud-bigquery
    #   pandas-gbq
pandas==2.3.1
    # via
    #   -r requirements.in
    #   db-dtypes
    #   pandas-gbq
pandas-gbq==0.23.1
    # via -r requirements.in
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-secret-manager
protobuf==4.25.8
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==16.1.0
    # via
    #   -r requirements.in
    #   db-dtypes
    #   pandas-gbq
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydata-google-auth==1.9.1
    # via pandas-gbq
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   brevo-python
    #   google-cloud-bigquery
    #   pandas
python-dotenv==1.1.1
    # via dotenv
pytz==2025.2
    # via pandas
requests==2.32.4
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==14.0.0
    # via typer
rsa==4.9.1
    # via google-auth
setuptools==80.9.0
    # via
    #   pandas-gbq
    #   pydata-google-auth
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   brevo-python
    #   python-dateutil
typer==0.16.0
    # via -r requirements.in
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   multidict
    #   rich
    #   typer
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   brevo-python
    #   requests
yarl==1.20.1
    # via aiohttp
